+++ thunar/thunar-shortcuts-view.c
@@ -405,7 +405,7 @@ thunar_shortcuts_view_button_release_event (GtkWidget      *widget,
   ThunarShortcutsView *view = THUNAR_SHORTCUTS_VIEW (widget);
 
   /* check if we have an event matching the pressed button state */
-  if (G_LIKELY (view->pressed_button == event->button))
+  if (G_LIKELY (view->pressed_button == (gint) event->button))
     {
       /* check if we should simply open or open in new window */
       if (G_LIKELY (event->button == 1))
